requirements.txt
anthropic
python-dotenv
requests
langgraph>=0.2.0
python-pptx
streamlit
